third_party/blink/web_tests/fast/forms/select/menulist-selection-reset.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<select id="select">
    <option>FAIL</option>
    <option>PASS</option>
    <option selected="true">Did not run</option>
</select>
<script>
test(() => {
    select.selectedIndex = 1;
    assert_equals(select.selectedIndex, 1);
});
</script>
